Simulacion de Partidos y Base de Datos desde LocalHost

+++ src/main/java/Core/domain/Stadium.java
@@ -1,18 +1,18 @@
 package Core.domain;
 
 public class Stadium {
-    private int id;
+    private long id;
     private String name;
     private City city;
 
-    public Stadium(int id, String name, City city) {
+    public Stadium(long id, String name, City city) {
         this.id = id;
         this.name = name;
         this.city = city;
     }
 
-    public int getId() { return id; }
-    public void setId(int id) { this.id = id; }
+    public long getId() { return id; }
+    public void setId(long id) { this.id = id; }
 
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
